Makes it possible to save and load account files
Account data can be saved with the AccountData.save(path) function.
Saved account data can be loaded just like accound data from banks
by using the account_file attribute in the constructor. Tag_file and such
data will need to be referenced when invocing the constructor as these are
not saved together with the account data.